Create 4 new routines in the module mod_oasis_print, to print the information in the pout files : - oasis_pprinti to print integers (3 max) - oasis_pprintr to print reals (3 max) - oasis_pprintc to print characters (3 max) - oasis_pprintl to print logicals (3 max) replace write(nulprt,*) with these routines
If prism_debug=0, only one file is written by the master proc of each model. Only the tree calling sequence informations are not printed in the pout files.
If prism_debug>0, all procs open and write in a pout file. The debug information printed in the file will depend of the value of nlogprt read in the namcouple (see ticket 235) 

For the file nout :
the unit of this file is now nulprt1.
Only the master process over all the processes of all the models, ie mpi_rank_global=0, opens and write the information contained in the namcouple in the nout file.


svn path=/trunk/oasis3-mct/; revision=97